requirements.txt
Flask>=3.0.0
Pillow>=10.0.0
pycryptodome>=3.20.0
gunicorn>=21.2.0
